BREAKING: update to v0.6.3 and remove custom entrypoint (#32)
This changes the behavior of how multiple GPUS were automatically detected. Previously we set tensor parallel to be equal to amount of GPUs. This isn't the correct approach since in some cases you want to use pipeline parallel across GPUs (e.g. on L4 GPUs).

In addition this allows people to easily switch between upstream official image and this image.